improvements

+++ vigpol/policer_config.c
@@ -19,7 +19,7 @@
 const uint32_t DEFAULT_LAN = 1;
 const uint32_t DEFAULT_WAN = 0;
 const uint64_t DEFAULT_RATE = 1000000; // 1MB/s
-const uint32_t DEFAULT_BURST = 100000; // 100kB
+const uint64_t DEFAULT_BURST = 100000; // 100kB
 const uint32_t DEFAULT_CAPACITY = 128; // IPs
 
 #define PARSE_ERROR(format, ...)                                               \
@@ -101,7 +101,7 @@ void nf_config_usage(void) {
           "\t--rate <rate>: policer rate in bytes/s,"
           " default: %" PRIu64 ".\n"
           "\t--burst <size>: policer burst size in bytes,"
-          " default: %" PRIu32 ".\n"
+          " default: %" PRIu64 ".\n"
           "\t--capacity <n>: policer table capacity,"
           " default: %" PRIu32 ".\n",
           DEFAULT_LAN, DEFAULT_WAN, DEFAULT_RATE, DEFAULT_BURST,
@@ -114,7 +114,7 @@ void nf_config_print(void) {
   NF_INFO("LAN Device: %" PRIu32, config.lan_device);
   NF_INFO("WAN Device: %" PRIu32, config.wan_device);
   NF_INFO("Rate: %" PRIu64, config.rate);
-  NF_INFO("Burst: %" PRIu32, config.burst);
+  NF_INFO("Burst: %" PRIu64, config.burst);
   NF_INFO("Capacity: %" PRIu16, config.dyn_capacity);
 
   NF_INFO("\n--- ------ ------ ---\n");
